Update Project.asm
Program collects five (2-digit) numbers and stores them in an array from addresses x3040 - x3044. If you search the addresses the numbers are in hex but if you convert the values to decimal, they will be equal to the values you inputted. Ex. If you use these values (DECIMAL - HEX)
(52 - 34)
(87 - 57)
(96 - 60)
(79 - 4F)
(61 - 3D)