Update Project.asm
Added a way to output the array. 
In the OUTLOOP the value is created but since two digit numbers cannot be outputted, I did a different method. I got the number and a divided it, and with the quotent and the remainder of the number, I displayed them differently. For example:
If we want to output a 14. 
14 / 10 = 1 Remainder: 4
x = 1
y = 4
OUT each variable separately.